+++ chat-widget/src/common/Constants.ts
@@ -114,6 +114,8 @@ export class Constants {
     public static readonly Title = "title";
     public static readonly Target = "target";
     public static readonly Blank = "_blank";
+    public static readonly TargetSelf = "_self";
+    public static readonly TargetTop = "_top";
     public static readonly TargetRelationship = "rel";
     public static readonly TargetRelationshipAttributes = "noopener noreferrer";