+++ chat-widget/src/components/webchatcontainerstateful/interfaces/IWebChatContainerStatefulProps.ts
@@ -13,6 +13,8 @@ export interface IWebChatContainerStatefulProps {
     containerStyles?: IStyle;
     disableNewLineMarkdownSupport?: boolean;
     disableMarkdownMessageFormatting?: boolean;
+    opensMarkdownLinksInSameTab?: boolean;
+    honorsTargetInHTMLLinks?: boolean;
     webChatStyles?: StyleOptions;
     webChatProps?: IWebChatProps;
     // eslint-disable-next-line @typescript-eslint/no-explicit-any